requirements.txt
streamlit
altair<5
numpy<2.0.0
langchain-core
langchain-community
langchain-groq
langchain-openai
langchain-huggingface
faiss-cpu
sentence-transformers>=3.0.0
PyPDF2
python-docx
python-pptx
torch
transformers<4.50.0
huggingface-hub>=0.20.0,<1.0.0
python-dotenv
fastapi
uvicorn[standard]
pydantic
sqlalchemy
pandas
requests
